+++ packages/core/src/decorators/index.ts
@@ -50,5 +50,4 @@ export * from "./UploadedFiles";
 export * from "./UseAfter";
 export * from "./UseBefore";
 export * from "./UseInterceptor";
-export * from "./Model";
 export * from "./Property";