+++ starters/openapi/src/decorator/index.ts
@@ -2,3 +2,4 @@ export {EnableOpenApi} from "./EnableOpenApi";
 export {EnableSwaggerUI} from "./EnableSwaggerUI";
 export {OpenAPI} from "./OpenAPI";
 export {ResponseSchema} from "./ResponseSchema";
+export {Model} from "./Model";