+++ starters/firebase/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nodeboot/starter-firebase
 
+## 1.0.4
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/core@1.3.0
+
 ## 1.0.3
 
 ### Patch Changes