+++ tests/core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nodeboot/test
 
+## 1.1.4
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/core@1.3.0
+
 ## 1.1.3
 
 ### Patch Changes