+++ tests/jest-fastify-demo/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @nodeboot/jest-fastify-demo
 
+## 1.0.6
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/starter-persistence@1.4.0
+    -   @nodeboot/starter-openapi@1.5.0
+    -   @nodeboot/core@1.3.0
+    -   @nodeboot/fastify-server@1.2.1
+
 ## 1.0.5
 
 ### Patch Changes