+++ tests/jest/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @nodeboot/jest
 
+## 1.1.4
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/core@1.3.0
+    -   @nodeboot/test@1.1.4
+
 ## 1.1.3
 
 ### Patch Changes